sys_fs: Lock dev_hdd1 mount point at cellSysCacheClear
